Added reversed cogstack module to prev version, added new module cogstack2 and added deprecation warning
